Refactored app.rb into several separate controllers. closes #78
